SHIT CLEAN
Former-commit-id: 05f083f08260c594e15c37818112703a1bc4f9e3

+++ srcs/cl_float3_manage.c
@@ -6,7 +6,7 @@
 /*   By: sbrella <sbrella@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/07 17:03:05 by sbrella           #+#    #+#             */
-/*   Updated: 2019/10/21 17:43:38 by sbrella          ###   ########.fr       */
+/*   Updated: 2019/10/21 18:35:21 by sbrella          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,8 +48,8 @@ cl_float3		rotate(cl_float3 axis, cl_float3 vector, float angle)
 	float		sin_a;
 	float		cos_a;
 
-	sin_a = sin(angle);
-	cos_a = cos(angle);
+	sin_a = (float)sin(angle);
+	cos_a = (float)cos(angle);
 	res.s[0] = (cos_a + (1 - cos_a) * axis.s[0] * axis.s[0]) * vector.s[0] +
 	((1 - cos_a) * axis.s[0] * axis.s[1] - sin_a * axis.s[2]) * vector.s[1] +
 	((1 - cos_a) * axis.s[0] * axis.s[2] + sin_a * axis.s[1]) * vector.s[2];
@@ -87,7 +87,7 @@ cl_float3	normalize(cl_float3 vector)
 	cl_float3	res;
 	float		norm;
 
-	norm = sqrt(vector.s[0] * vector.s[0] + vector.s[1] * vector.s[1] + vector.s[2] * vector.s[2]);
+	norm = (float)sqrt(vector.s[0] * vector.s[0] + vector.s[1] * vector.s[1] + vector.s[2] * vector.s[2]);
 	res.s[0] = vector.s[0] / norm;
 	res.s[1] = vector.s[1] / norm;
 	res.s[2] = vector.s[2] / norm;
